moved config files to config package changed thermalconfig to read from config.toml and lock
