+++ classify/main.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from .clipclassifier import ClipClassifier
 from ml_tools.logs import init_logging
 from ml_tools import tools
-from ml_tools.config import Config
+from config.config import Config
 from ml_tools.previewer import Previewer
 import absl.logging
