+++ requirements.txt
@@ -15,3 +15,5 @@ attrs~=19.1
 filelock~=3.0.12
 Astral~=1.10.1
 timezonefinder~=4.1.0
+psutil~=5.6.7
+portalocker~=1.5.2
